Modify as reviewer commented
