+++ paddle/fluid/framework/paddle2cinn/cinn_compiled_object.h
@@ -12,30 +12,39 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gtest/gtest.h"
+#pragma once
 
-#include "paddle/fluid/framework/cinn/cinn_runner.h"
+#include <map>
+
+#include "paddle/fluid/framework/feed_fetch_type.h"
 #include "paddle/fluid/framework/ir/graph.h"
 #include "paddle/fluid/framework/lod_tensor.h"
-#include "paddle/fluid/framework/program_desc.h"
 #include "paddle/fluid/framework/scope.h"
 
 namespace paddle {
 namespace framework {
-namespace cinn {
+namespace paddle2cinn {
 
-using ir::Graph;
+// Class to store and call CINN complied object
+class CinnCompiledObject {
+ public:
+  CinnCompiledObject();
+  ~CinnCompiledObject();
 
-TEST(CinnRunnerTest, TodoTest) {
-  ProgramDesc empty_program;
-  Graph empty_graph(empty_program);
-  Scope empty_scope;
-  std::map<std::string, const LoDTensor*> empty_feed;
+  // Compiles use CINN. CINN compilation needs model graph, input names, and
+  // input_shapes
+  void Compile(const ir::Graph& graph,
+               std::map<std::string, const LoDTensor*>* feed_targets);
 
-  CinnRunner cinn_runner;
-  cinn_runner.Run(empty_graph, &empty_scope, &empty_feed);
-}
+  // Feed LoDTensors to tun CINN compiled object and return fetched result
+  std::map<std::string, FetchType*> Run(
+      Scope* scope, std::map<std::string, const LoDTensor*>* feed_targets);
 
-}  // namespace cinn
+  // Converts compiled object to Paddle Graph
+  // To be discussed
+  // ir::Graph ToGraph();
+};
+
+}  // namespace paddle2cinn
 }  // namespace framework
 }  // namespace paddle
